[Silver I] Title: 포도주 시식, Time: 84 ms, Memory: 12504 KB -BaekjoonHub

+++ 백준/Silver/2156. 포도주 시식/포도주 시식.java
@@ -8,8 +8,8 @@ class Main{
         int n = Integer.parseInt(br.readLine());
 
 
-        long[] dp = new long[n + 1];
-        long[] arr = new long[n + 1];
+        int[] dp = new int[n + 1];
+        int[] arr = new int[n + 1];
         for(int i = 1; i <= n; i++) arr[i] = Integer.parseInt(br.readLine());
